network: refactoring to use the new structure
